Added Iterate 6
Added the CheckLevelState to the Level.cs file. Made use of the IsIndexValid method that would check if there are elements within the grid.
If this was false, then I would change the LevelState to NoElementsLeft

I also used the GetAdjacentCellsOfSameType method and if this was smaller than 2 then I would set the LevelState to NoMoreMovePossible